refactor: group scan monitoring components into scan/monitor/
