+++ src/components/scan/monitor/ScanCompletionSync.tsx
