+++ src/components/scan/monitor/scan-progress-bar.tsx
